float_to_string::without_valid_option_errors_badarg
